+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/implementation/query/DocumentQueryExecutionContextBase.java
@@ -281,6 +281,10 @@ implements IDocumentQueryExecutionContext<T> {
                 requestHeaders.put(HttpConstants.HttpHeaders.DEDICATED_GATEWAY_PER_REQUEST_BYPASS_CACHE,
                     String.valueOf(cosmosQueryRequestOptions.getDedicatedGatewayRequestOptions().isIntegratedCacheBypassed()));
             }
+            if (cosmosQueryRequestOptions.getDedicatedGatewayRequestOptions().getShardKey() != null) {
+                requestHeaders.put(HttpConstants.HttpHeaders.DEDICATED_GATEWAY_PER_REQUEST_SHARD_KEY,
+                    cosmosQueryRequestOptions.getDedicatedGatewayRequestOptions().getShardKey());
+            }
         }
 
         if (cosmosQueryRequestOptions.isIndexMetricsEnabled()) {